CallbackRunnable: Move to "components" namespace
